requirements.txt
torch>=1.1
torchvision~=0.11.2
numpy~=1.21.5
tqdm~=4.62.3
tensorboard>=1.14

thop~=0.0.31.post2005241907
requests~=2.25.1
Pillow~=8.4.0
matplotlib~=3.5.1
opencv-python~=4.5.4.60
PyYAML~=5.3.1
scipy~=1.7.3
pandas~=1.3.5
seaborn~=0.11.2